+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Added
 
 - .zshrc-ci file in `.impaktfull/impaktfull_cli` when running `ci_cd setup`
+- Added support for getting the github build nr from the environment variables (GITHUB_RUN_NUMBER)
 
 # 0.27.2